+++ src/Instance/ItemsList/utils.js
@@ -6,6 +6,7 @@ const sorters = {
   'barcode': ({ barcode }) => barcode,
   'status': ({ status }) => status.name.toLowerCase(),
   'materialType': ({ materialType }) => materialType.name.toLowerCase(),
+  'effectiveLocation': ({ effectiveLocation }) => effectiveLocation.name.toLowerCase(),
   'enumeration': ({ enumeration }) => enumeration,
   'chronology': ({ chronology }) => chronology,
   'volume': ({ volume }) => volume,